tustus_2.7.4a/requirements.txt
python-telegram-bot[job-queue]==21.6
requests
beautifulsoup4
lxml
tenacity
python-dotenv
